+++ package.json
@@ -13,6 +13,7 @@
     "@react-native-firebase/app": "^22.4.0",
     "@react-native-firebase/auth": "^22.4.0",
     "@react-native/new-app-screen": "0.80.2",
+    "@react-navigation/bottom-tabs": "^7.4.5",
     "@react-navigation/elements": "^2.6.2",
     "@react-navigation/native": "^7.1.17",
     "@react-navigation/native-stack": "^7.3.24",
