+++ src/componets/AuthFormComponent.tsx
@@ -107,21 +107,6 @@ const AuthFormComponent = (props: AuthFormComponentProps) => {
 };
 
 const styles = StyleSheet.create({
-  bgImage: {
-    flex: 1,
-  },
-  container: {
-    flex: 1,
-    backgroundColor: '#0524336c',
-    paddingHorizontal: 13,
-    paddingVertical: 24,
-    justifyContent: 'space-between',
-  },
-  logo: {
-    width: 160,
-    height: 48,
-    resizeMode: 'contain',
-  },
   formContainer: {
     backgroundColor: '#ffffff69',
     borderRadius: 40,
