+++ src/navigation/AuthStack.tsx
@@ -1,8 +1,8 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { RootStackParamList } from './type';
-import LoginScreen from '../screens/LoginScreen';
-import SignupScreen from '../screens/SignupScreen';
-import HomeScreen from '../screens/HomeScreen';
+import LoginScreen from '../screens/auth/LoginScreen';
+import SignupScreen from '../screens/auth/SignupScreen';
+import HomeScreen from '../screens/bottom-tabs/HomeScreen';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
