+++ src/screens/auth/style/styles.tsx
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  bgImage: {
+    flex: 1,
+  },
+  container: {
+    flex: 1,
+    backgroundColor: '#0524336c',
+    paddingHorizontal: 13,
+    paddingVertical: 24,
+    justifyContent: 'space-between',
+  },
+  logo: {
+    width: 160,
+    height: 48,
+    resizeMode: 'contain',
+  },
+});
+
+export default styles;
